add redirect from feed router in the home page
