+++ src/main.tsx
@@ -10,6 +10,7 @@ const router = createBrowserRouter([
   {
     path: "/",
     element: <Layout />,
+
     children: [
       {
         path: "/feed",
