+++ src/styles/GlobalStyle.ts
@@ -3,4 +3,8 @@ import { createGlobalStyle } from "styled-components";
 export const GlobalStyle = createGlobalStyle`
 body{
     font-family:"Roboto";
-}`;
+    background-color: #343434;
+}
+
+
+`;
